% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-naming.R
\name{hierarchy_table_name}
\alias{hierarchy_table_name}
\title{Construct the standardized table name for a cell hierarchy table}
\usage{
hierarchy_table_name(fine_arcmin, coarse_arcmin)
}
\arguments{
\item{fine_arcmin}{Numeric resolution of the finer grid in arcminutes}

\item{coarse_arcmin}{Numeric resolution of the coarser grid in arcminutes}
}
\value{
Character table name, e.g. "cell_hierarchy_0_5_to_15_arcmin"
}
\description{
Construct the standardized table name for a cell hierarchy table
}
