% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-boundaries.R
\name{resolve_admin_id}
\alias{resolve_admin_id}
\title{Resolve an admin_name to its admin_id within a given admin_level}
\usage{
resolve_admin_id(con, admin_level, admin_name)
}
\arguments{
\item{con}{A DBI connection}

\item{admin_level}{Character, e.g. "ADM1"}

\item{admin_name}{Character, human-readable name (case-insensitive match)}
}
\value{
Character admin_id
}
\description{
Resolve an admin_name to its admin_id within a given admin_level
}
